crush/create_or_move: Adjust weight of destroyed item if different
If an OSD is restarted/re-added with a different weight we should
adjust the CRUSH weight of the item since it could be a different
weight then it was before.

An OSD might be destroyed and re-added with a differently sized
medium (SSD / HDD) underneath and we should adjust the weight
accordingly.

Fixes: http://tracker.ceph.com/issues/23467

Signed-off-by: Wido den Hollander <wido@42on.com>