osd: do not release_reserved_pushes when requeuing
Back in 3cc48278bf0ee5c9535d04b60a661f988c50063b we refactoring the sharded
wq and incorrectly included code that would release_reserved_pushes for
items that were queued and deferred and then woken and put back in the
queue. The reserved_pushes are for recovery ops that are in flight in the
queue, which includes the priority queue *and* the waiting_for_pg; the code
we replaced would release these only when dequeueing an item (or items) for
processing (or discard).

In master, this code is fixed as part of the peering fast dispatch and
OSDShard refactor.

Signed-off-by: Sage Weil <sage@redhat.com>