osd: resend osd_pgtemp if it's not acked
if the osd_pgtemp message is dropped before monitor receives it, we need
to resend it. otherwise a pg could be stuck in activating state if the
pg creation was withheld by the max-pg-per-osd on the replica, and then
the replica osd removes some existing pg.

Fixes: http://tracker.ceph.com/issues/23610
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 9c5cb4ea620da4ef14345ebc8018f9ed008c4345)

Conflicts:
	src/osd/OSD.cc: structured binding is a feature introduced in
C++17. while we are using C++11 in luminous. so implement it with C++11
syntax.
	src/osd/OSD.h: trivial resolution