install/rpm: add install_ceph_packages and repos_only options
This commit adds two new boolean options to the install task:

install_ceph_packages (defaults to "True") - controls whether or not ceph
packages are installed. Note: the "librados2" and "ceph-test" packages get
installed even if this option is set to False.

repos_only (defaults to "False") - when this option is set to "True",
the install task merely adds repos and does nothing else.

The new options currently work only on RPM systems: on DEB systems they are
ignored.

Signed-off-by: Nathan Cutler <ncutler@suse.com>