mgr/dashboard: Change deletion link to modal only
Due to CSS problems the link solution wasn't the best way, now it will
represent only the modal content. The downside of this solution is that
it put's the burden on the developer to use it the right way and import
a view things to get it working. But on the upside CSS styles will work
as expected.

The unit test example was updated accordingly this way it should be easy
to understand how it can be implemented the right way.

Signed-off-by: Stephan Müller <smueller@suse.com>
(cherry picked from commit 63ae858582ec48487e28acf59a4dc315bd423303)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>