mgr: introduce MgrModule.OPTIONS field
Now is a good time to start requiring
modules to explicitly list their configuration
settings, so that we can do a proper job of
migrating configuration from old config-key style,
i.e. knowing what's a config setting and what's
a KV store item.

Throw an exception if a module tries to
access a setting outside their schema, so
that we have some confidence that the schema
is complete.

Signed-off-by: John Spray <john.spray@redhat.com>