mgr/dashboard: restcontroller: support explicit resource ID
In the current implementation, RESTController tries to infer the
resource ID by looking at the declared parameters of the resource
methods (get, delete, set, and patch). This inference may not always
be correct, and therefore with this commit we can now explicitely
specify the structure of the resource ID. If the resource ID is
not explicitly specified, then we fallback to the inference algorithm.

Signed-off-by: Ricardo Dias <rdias@suse.com>