ceph-detect-init: stop using platform.linux_distribution
This commit introduces a new function, _extract_from_os_release, which is
used (only on Linux) to extract the ID and VERSION_ID fields from the
/etc/os-release file. This is sufficient to positively determine the OS
and version on these systems, and enable ceph-detect-init to correctly
return systemd on them.

The commit includes unit tests asserting that _extract_from_os_release really
extracts the strings that we expect it to, using the contents of
/etc/os-release from a number of different operating systems targeted by Ceph.

This commit is not cherry-picked because ceph-detect-init has been dropped in
master.

Fixes: http://tracker.ceph.com/issues/18163
Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit 12d94cae2a06fd2ae39379e233588be99bff91de)