Fix broken SIMD tests on ARM
The existing feature tests were checking if the compiler is capable of
generating code with given features, not if the target architecture
specificed by march/mcpu supports the features. This lead to Ceph
crashing due to illegal instructions.

This patch replaces the tests with actual checks for the target
architecture, and removes any explicit setting of march since that is
now inferred from the environment.

Signed-off-by: Robin McCorkell <robin@mccorkell.me.uk>