MDSMonitor: clean up use of pending fsmap in uncommitted ops
Gist of this commit is to eliminate use of PaxosFSMap::get_working_fsmap which
allowed looking at the pending FSMap depending if it's the leader. Instead,
pass the FSMap being worked on to helper functions.

Also, ensure that we are only looking at the committed (i.e. current) FSMap
in the preprocess_* service methods.

Fixes: http://tracker.ceph.com/issues/23768

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>