mgr/dashboard: Format small numbers correctly
The issue was triggered by numbers that a lower than 1.
Doing a logarithm with a number lower than 1 leads to
negative value that is not handled anywhere in The formatter service as
a result the final value will be quirky.
The negative number will also be used as index in the units array, where
it will return "undefined".

Fixes #24081

Signed-off-by: Stephan Müller <smueller@suse.com>
(cherry picked from commit b8ecac3603c01764d83e2b8862f3b6f847321e85)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>