mgr_module: Deal with long running avgs properly
We have recently started exposing long running avgs with the mgr python
modules but that only covers the total sum, not the avgcount.

With this patch, the counts for long running avgs are exposed to the
python modules allowing them to deal with the new data.

Signed-off-by: Boris Ranto <branto@redhat.com>
(cherry picked from commit f79ff11302a0364c04601b5ccd3950d80cd063a8)