mgr/dashboard: Fix Jest conflict with coverage files
It seems 'jest --watch' is trying to determine the modules that are imported
inside the coverage folder.
Since this folder is not always up to date, when there are references to
nonexistent files, Jest will simply break and hang still.

I added an explicit rule for Jest to ignore this folder.

Signed-off-by: Tiago Melo <tmelo@suse.com>