ceph-fuse: Delete inode's bufferhead was in Tx state would lead a assert fail
Prematurely delted the bh which was in Tx state may lead the object can't
be closed before its writer of this bh callback. Thus if inode's ref call
put_inode decrease ref to zero and release inode's oset.An assert fail occur
beacuse the oset can't be emptied.

If we did discard an object, when write callback of one bh belong to the
object and found the oset->dirty_or_tx become to 0, we need skip it in
case of double dropping of caps.