mon/MonMap: separate rank ordering from entity_addr_t
We currently define the mon rank ordering based on the sort order of the
mon addresses.  Change that so that the rank order is explicitly encoded
in the ranks field of the MonMap.

If we load a legacy MonMap, calculate the legacy ordering.

If the monmap does not require nautilus features yet, force the legacy
ordering.  Once all mons are >= nautilus, we can reorder ranks.  Note that
the daemons and clients (MonClients) may see a different rank ordering.
That should be okay.

Signed-off-by: Sage Weil <sage@redhat.com>