common/Throttle: extract ThrottleInterface
the Message classes are shared by OSD and other components of Ceph,
and the throttle in Policy class is different in seastar and
non-seastar world. we will have different implementations for the
seastar applications and non-seastar apps, to consolidate these
two implementations, we need to introduce a common interface for
them.

Signed-off-by: Kefu Chai <kchai@redhat.com>