dashboard: Return float if rate not available
This fixes a test case issue when some of the data might not be
available during the test, yet. The test expects the second value to be
JFloat but that fails because get_rates return an int in that case.

Signed-off-by: Boris Ranto <branto@redhat.com>
(cherry picked from commit 58ffd146261586b5598c946fd730df0b223294e7)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>