rgw: fetch_remote_obj() applies olh even if object is current
even if we already have the latest version of an object, it may have
synced without a valid olh epoch. so we still need to apply its olh
modifications with this epoch. this change also has to be added to the
data changes log

Signed-off-by: Casey Bodley <cbodley@redhat.com>
(cherry picked from commit b5ecc4eaaaaa4f9950ad934a015255171abeb01e)