ceph-volume lvm.activate Do not search for a MON configuration
This system might not have one nor can we be sure that we have
a client.admin keyring on the system.

Just prime the directory and have the OSD then use it's own cephx
key to communicate with the MONs.

Signed-off-by: Wido den Hollander <wido@42on.com>
(cherry picked from commit 722d79d16a0da042bbc189709efd3eeff1366558)