mgr/dashboard: Listen on port 7000 by default and not 8080
Port 8080 is a common alternative HTTP port used for web traffic.

The dashboard however uses SSL (which can not be turned off) and for
that purpose we should either use 8443.

Historically the dashboard would run on port 7000 so to keep this backwards
compatibility the dashboard should bind again on TCP 7000.

Signed-off-by: Wido den Hollander <wido@42on.com>