mgr/telegraf: Catch OSError instead of errno
A OSError will be thrown (and not catched) if the module starts for
the first time and the UNIX socket of the Telegraf Agent does not
exist (yet).

This will render the module useless because it can not accept commands
for configuration at that point.

If we catch an exception, simply log it but do not raise it.

Signed-off-by: Wido den Hollander <wido@42on.com>