osd/PG: reset PG peering if osd transitions from down -> up
Consider a PG that is stray and ends up in ReplicaActive (because it is
participating as a recovery source).  If it is marked down wrongly and
then comes back up, then the PG will not reset, because there was not
an interval change (the PG is not part of the up or acting sets).
This can leave the PG in an odd state, leading to questionable behavior.
(For example, a stray might be in ReplicaActive and then ignore some
types of query messages.)

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 8d3b4ca591090a5ab3b5940f43a86866f2f5b9bc)