ctest: Removing unittest_alloc_bench
unittest_alloc_bench is very cpu consuming and can take up to 20mn to
run. As per a discussion with the original author of this code, this
test is only about testing the performance while there is no validation
code in it.

To save time from people doing make check often or even the CI itself,
removing this test will save a lot of time while not reducing the test
coverage.

This commit is only removing the test from the make check but keep the
binary being compiled for those who want to run it manually.

Signed-off-by: Erwan Velu <erwan@redhat.com>