osd/PG: do not assume delete event means found_missing
This condition was introduced in 5ad79625d65ca8ff7d55040124c483efa8bd4a88
as part of the missing deletes series, without a clear motivation.  The
best guess is that it was either compensating for some other unfound bug
or simply being a bit overaggressive.

The problem is that it triggers a notify being sent to the sender, or
restarts recovery, both of which are overreactions.

Signed-off-by: Sage Weil <sage@redhat.com>