doc: Mention PURGED_SNAPDIRS and RECOVERY_DELETES in Mimic release notes
71d1dcbb98bf1b3c7953c13a1165ba1cac7e7fd0 started enforcing the
PURGED_SNAPDIRS osdmap flag before am upgraded Mon could join the
quorum.

9b80b14783ef895390b4153320078661627f373d extended that check to also
include RECOVERY_DELETES. The Mimic release notes do not mention these
flags as prerequisites for an upgrade beyond Luminous. That creates an
obvious issue for users who skipped Luminous in production, and now
want to upgrade from Jewel to Mimic in, say, a weekend.

Update the release notes to include those flags as prerequisites for a
Luminous to Mimic upgrade, explain how users can make sure that they
are set, and also give users a one-liner to fix up their PGs in a
pinch, if they need to.

Signed-off-by: Florian Haas <florian@citynetwork.eu>