osd: make build_past_intervals_parallel() optional, and off by default
The implementation is not very tolerant and can cause problems.  For
example, with bug http://tracker.ceph.com/issues/21142 the pg epoch_created
was off, leading to an attempt to rebuild intevals prior to the PG's.
Allow it to be skipped entirely.  (Note that this method is removed
entirely in mimic, see ac9e2f76093ef00238a1a8a0266a95c5ca2468b2).

Signed-off-by: Sage Weil <sage@redhat.com>