mgr/dashboard: Improve SummaryService and TaskWrapperService
When you create a new task, and it stays running, it will be added automatically
to the summary data.

This will allows for us to deal with it more quickly, by subscribing
to the summaryService, and removes the need to pass a runningTasks array
between services.

Added 3 new methods to SummaryService.

Signed-off-by: Tiago Melo <tmelo@suse.com>
(cherry picked from commit 5ee689feb6d4dfcf8949cc8f79c84d6ddeb81a31)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>