src/ceph-disk: Reintroducing ceph-disk
PR https://github.com/ceph/ceph/pull/22343 have been removing ceph-disk
as ceph-volume is the replacement tool.

As long as ceph-volume is not able to detect all the ceph-disk
configurations, we need to keep ceph-disk for restoring the "list"
feature.

This commit is the first of a series that will restore the minimal
subset to get "ceph-disk list" working.

ceph-detect-init & tests will not be restored.

Signed-off-by: Erwan Velu <erwan@redhat.com>